// src/types/event.ts
export interface Event {
    _id: string;
    userId: string;
    eventTitle: string;
    eventVenue: string;
    eventStartDate: string;
    eventEndDate: string;
    eventStartTime: string;
    eventEndTime: string;
    eventCoverCost: number;
    eventImage: string;
    eventDescription: string;
    createdAt: string;
    updatedAt: string;
    __v: number;
  }
  